2_DeployModel/requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
Flask==2.2.2
idna==3.3
importlib-metadata==5.0.0
ipykernel==6.13.0
ipython==8.3.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.0
jupyter-core==4.10.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
nltk==3.7
numpy==1.23.1
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
pyzmq==22.3.0
regex==2022.7.25
requests==2.28.1
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
stack-data==0.2.0
torch==1.12.1
torchvision==0.13.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.8.1
